projects/plywood-chatbot/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
langchain>=0.0.300
openai>=1.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
pydantic>=1.10.7
PyPDF2>=3.0.0
rich>=13.0.0